Refactor toolbar into BaseActivity
