+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">Stardew Valley Pocket Tracker</string>
+    <string name="activity_main">Home</string>
+    <string name="activity_settings">Settings</string>
 </resources>
